:art: move and refact OrderStatusEnum to Enum directory
