+++ catalog/model/extension/payment/mundipagg_order_processing.php
@@ -1,6 +1,6 @@
 <?php
 
-use Mundipagg\OrderstatusEnum;
+use Mundipagg\Enum\OrderstatusEnum;
 
 class ModelExtensionPaymentMundipaggOrderProcessing extends Model
 {
